Humana downloaded
One error exists. For some reason the else statement in the while loop never runs. Even when the if statement conditions are not met, the else still doesn't run and cancel the loop. Fortunately, the if doesn't run so only R is running and not the connection to the server.